Fixed build

+++ skuratov.danil/P4/countSaddlePoints.cpp
@@ -2,7 +2,7 @@
 
 #include <cstddef>
 
-int skuratov::countSaddlePoints(int* array, size_t rows, size_t cols)
+size_t skuratov::countSaddlePoints(int* array, size_t rows, size_t cols)
 {
   size_t saddleCount = 0;
   for (size_t i = 0; i < rows; ++i)
