+++ skuratov.danil/P4/countSaddlePoints.h
@@ -5,7 +5,7 @@
 
 namespace skuratov
 {
-  int countSaddlePoints(int* array, size_t rows, size_t cols);
+  size_t countSaddlePoints(int* array, size_t rows, size_t cols);
 }
 
 #endif
