+++ skuratov.danil/P4/findLongestSeriesRow.cpp
@@ -2,11 +2,11 @@
 
 #include <cstddef>
 
-int skuratov::findLongestSeriesRow(const int* array, size_t rows, size_t cols)
+size_t skuratov::findLongestSeriesRow(const int* array, size_t rows, size_t cols)
 {
-  int maxSeriesLength = 1;
-  int currentSeriesLength = 1;
-  int maxSeriesRow = 0;
+  size_t maxSeriesLength = 1;
+  size_t currentSeriesLength = 1;
+  size_t maxSeriesRow = 0;
   for (size_t i = 0; i < rows; ++i)
   {
     currentSeriesLength = 1;
